pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "fpdf>=1.7.2",
    "gunicorn>=23.0.0",
    "langchain>=0.3.23",
    "langchain-anthropic>=0.3.10",
    "langchain-core>=0.3.51",
    "langchain-openai>=0.3.12",
    "openai>=1.70.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
]
